# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FOODORO POS & Kitchen Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /waste:
    get:
      operationId: listWasteLogs
      tags: [waste]
      summary: List waste log entries
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: inventoryId
          in: query
          schema:
            type: integer
        - name: reason
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of waste entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteLog"
    post:
      operationId: createWasteLog
      tags: [waste]
      summary: Log a waste entry from the kitchen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWasteLogBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteLog"
  /waste/analytics:
    get:
      operationId: getWasteAnalytics
      tags: [waste]
      summary: Waste analytics — top wasted items, cost, trends
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Waste analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteAnalytics"

  # ─── Public (no auth) ─────────────────────────────────────────────────────
...
  /tables:
    get:
      operationId: listTables
      tags: [orders]
      summary: List restaurant tables
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantTable"
    post:
      operationId: createTable
      tags: [admin]
      summary: Create restaurant table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantTableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantTable"
  /tables/{id}:
    patch:
      operationId: updateTable
      tags: [admin]
      summary: Update table
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantTableInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantTable"
    delete:
      operationId: deleteTable
      tags: [admin]
      summary: Delete table
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
  /tables/{id}/seat:
    post:
      operationId: seatTable
      tags: [orders]
      summary: Seat customers at table
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerName: { type: string }
                guestCount: { type: integer }
      responses:
        "200":
          description: Updated
  /tables/{id}/clear:
    post:
      operationId: clearTable
      tags: [orders]
      summary: Clear/free a table
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                needsCleaning: { type: boolean }
      responses:
        "200":
          description: Cleared
  /reservations:
    get:
      operationId: listReservations
      tags: [orders]
      summary: List table reservations
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [orders]
      summary: Create reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created
  /reservations/{id}:
    patch:
      operationId: updateReservation
      tags: [orders]
      summary: Update reservation status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string }
      responses:
        "200":
          description: Updated

  # ─── Security ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RestaurantTable:
      type: object
      required: [id, number, capacity, status, posX, posY, shape, section, isActive]
      properties:
        id: { type: integer }
        number: { type: string }
        capacity: { type: integer }
        status: { type: string }
        posX: { type: integer }
        posY: { type: integer }
        shape: { type: string }
        section: { type: string }
        isActive: { type: boolean }
        currentOrderId: { type: ["integer", "null"] }
        occupiedSince: { type: ["string", "null"], format: date-time }
        customerName: { type: ["string", "null"] }
        guestCount: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    RestaurantTableInput:
      type: object
      required: [number]
      properties:
        number: { type: string }
        capacity: { type: integer }
        posX: { type: integer }
        posY: { type: integer }
        shape: { type: string }
        section: { type: string }
        status: { type: string }
        customerName: { type: string }
        guestCount: { type: integer }
        isActive: { type: boolean }
    Reservation:
      type: object
      required: [id, tableId, customerName, guestCount, reservationTime, status]
      properties:
        id: { type: integer }
        tableId: { type: integer }
        tableNumber: { type: ["string", "null"] }
        customerName: { type: string }
        customerPhone: { type: ["string", "null"] }
        guestCount: { type: integer }
        reservationTime: { type: string, format: date-time }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ReservationInput:
      type: object
      required: [tableId, customerName, reservationTime]
      properties:
        tableId: { type: integer }
        customerName: { type: string }
        customerPhone: { type: string }
        guestCount: { type: integer }
        reservationTime: { type: string, format: date-time }
        notes: { type: string }

    # Kitchen Availability
...
    WasteLog:
      type: object
      required: [id, tenantId, inventoryName, quantity, unit, reason, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        inventoryId:
          type: ["integer", "null"]
        inventoryName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        reason:
          type: string
          enum: [spoilage, burning, expiry, prep_error, theft, other]
        notes:
          type: ["string", "null"]
        loggedBy:
          type: ["string", "null"]
        costEstimate:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    CreateWasteLogBody:
      type: object
      required: [inventoryName, quantity, unit, reason]
      properties:
        inventoryId:
          type: ["integer", "null"]
        inventoryName:
          type: string
        quantity:
          type: number
          minimum: 0.0001
        unit:
          type: string
        reason:
          type: string
          enum: [spoilage, burning, expiry, prep_error, theft, other]
        notes:
          type: ["string", "null"]
        loggedBy:
          type: ["string", "null"]
        costEstimate:
          type: ["number", "null"]
        deductFromInventory:
          type: boolean
          default: true
    WasteAnalyticsItem:
      type: object
      required: [inventoryId, inventoryName, unit, totalWasted, totalCost, count]
      properties:
        inventoryId:
          type: ["integer", "null"]
        inventoryName:
          type: string
        unit:
          type: string
        totalWasted:
          type: number
        totalCost:
          type: number
        count:
          type: integer
    WasteAnalytics:
      type: object
      required: [totalWasteCost, totalEntries, topWastedItems, byReason, dailyTrend]
      properties:
        totalWasteCost:
          type: number
        totalEntries:
          type: integer
        topWastedItems:
          type: array
          items:
            $ref: "#/components/schemas/WasteAnalyticsItem"
        byReason:
          type: object
          additionalProperties:
            type: number
        dailyTrend:
          type: array
          items:
            type: object
            required: [date, totalCost, entries]
            properties:
              date:
                type: string
              totalCost:
                type: number
              entries:
                type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT